backend/requirements.txt
python-dotenv==1.0.1
Flask
Flask-SQLAlchemy
Flask-Migrate
mariadb
Flask-Cors==5.0.10

requests
pandas
dotenv
sympy
datetime

requests
dataclasses